custom dfu-util for commercial and alliance workshops
issue with current dfu-util is that at the end of flashing
arduino 101 it gives error while resetting it and the
imraa.lock file does not get created. Although the firmware
is flashed the lock file is not created which can be an
issue.

This version of dfu-util does a reset first followed by a
detach which resolves the issue - the original dfu-util has
reverse sequence which does not seem to work